+++ src/lib/security/csrf-protection.ts
@@ -1,34 +1,34 @@
 /**
- * CSRF保護機能（サーバーサイド専用）
+ * CSRF保護機能（Edge Runtime / Node.js 両対応）
  * Cross-Site Request Forgery攻撃を防止するためのトークン管理
  *
- * ⚠️ 【重要警告】: このファイルは絶対にクライアントサイドからインポートしないこと！
- * - API routes、Server Actions、Middlewareでのみ使用可能
- * - node:cryptoがブラウザで動作しないため、クライアントサイドでバンドルするとエラーになる
- * - クライアントサイドではCSRF Providerとuse-csrf-tokenフックを使用すること
- *
- * 使用可能場所：
- * ✅ src/app/api/*.ts (API routes)
- * ✅ src/middleware.ts
+ * Web Crypto API（`globalThis.crypto`）を使用しているため、以下すべてで利用可能：
+ * ✅ API routes (src/app/api/*)
+ * ✅ Middleware (src/middleware.ts, Edge Runtime)
  * ✅ Server Actions (_actions/admin/*.ts)
- * ❌ コンポーネント、フック、クライアント側プロバイダー
+ *
+ * クライアントサイドからは絶対にインポートしないこと（CSRF_SECRETがバンドルに含まれてしまうため）。
+ * クライアントではCSRF Providerとuse-csrf-tokenフックを使用すること。
+ *
+ * Double-submit Cookie パターン:
+ *   - `csrf-token` Cookie と `X-CSRF-Token` ヘッダーの両方が必須かつ値一致を要求
+ *   - クライアントは `/api/csrf-token` のJSONレスポンスからトークンを取得し、ヘッダーで送信
+ *   - CookieはHttpOnlyで発行され、JavaScriptから読めない
  */
 
-import { createHash, randomBytes } from "node:crypto";
 import type { NextRequest } from "next/server";
-import logger from "@/lib/logger";
 
 /**
  * 起動時に `CSRF_SECRET` 環境変数を検証して返す。
  *
  * - 必須環境変数。未設定または32文字未満の場合はエラーをスローし、
- *   Node.jsプロセス起動を停止する（フェイルクローズド）。
+ *   プロセス起動を停止する（フェイルクローズド）。
  * - 推奨生成方法: `openssl rand -hex 32`
  *
  * @throws {Error} `CSRF_SECRET` 未設定 / 32文字未満のとき
  * @returns 検証済みのCSRF秘密鍵
  */
-function getCSRFSecret(): string {
+export function getCSRFSecret(): string {
 	const secret = process.env.CSRF_SECRET;
 	if (!secret || secret.length < 32) {
 		throw new Error(
@@ -42,31 +42,52 @@ function getCSRFSecret(): string {
 const CSRF_SECRET = getCSRFSecret();
 
 /**
- * CSRFトークンを生成（サーバーサイド専用）
- * タイムスタンプと署名を含む安全なトークンを作成
+ * Web Crypto APIでSHA-256ハッシュを生成し、16進文字列で返す。
+ *
+ * @param data ハッシュ化する文字列
+ * @returns SHA-256ハッシュの16進表現（64文字）
+ */
+async function createSha256Hash(data: string): Promise<string> {
+	const encoder = new TextEncoder();
+	const dataBuffer = encoder.encode(data);
+	const hashBuffer = await crypto.subtle.digest("SHA-256", dataBuffer);
+	const hashArray = Array.from(new Uint8Array(hashBuffer));
+	return hashArray.map((b) => b.toString(16).padStart(2, "0")).join("");
+}
+
+/**
+ * ランダムな32バイトトークンを16進で生成する（Web Crypto API使用）。
+ *
+ * @returns 64文字の16進ランダム文字列
+ */
+function generateRandomToken(): string {
+	const array = new Uint8Array(32);
+	crypto.getRandomValues(array);
+	return Array.from(array, (byte) => byte.toString(16).padStart(2, "0")).join("");
+}
+
+/**
+ * CSRFトークンを生成する。
+ * フォーマット: `<random>:<timestamp>:<sha256-signature>`
  *
  * @returns 署名付きCSRFトークン
- * @serverOnly このファイルはサーバーサイドでのみ使用可能
  */
-export function generateCSRFToken(): string {
-	const token = randomBytes(32).toString("hex");
+export async function generateCSRFToken(): Promise<string> {
+	const token = generateRandomToken();
 	const timestamp = Date.now().toString();
-	const signature = createHash("sha256")
-		.update(`${token}:${timestamp}:${CSRF_SECRET}`)
-		.digest("hex");
+	const signature = await createSha256Hash(`${token}:${timestamp}:${CSRF_SECRET}`);
 
 	return `${token}:${timestamp}:${signature}`;
 }
 
 /**
- * CSRFトークンを検証（サーバーサイド専用）
- * タイムスタンプと署名の整合性をチェック
+ * CSRFトークンを検証する。
+ * タイムスタンプ（1時間以内）と署名の整合性をチェック。
  *
  * @param token 検証するCSRFトークン
  * @returns トークンが有効かどうか
- * @serverOnly このファイルはサーバーサイドでのみ使用可能
  */
-export function verifyCSRFToken(token: string): boolean {
+export async function verifyCSRFToken(token: string): Promise<boolean> {
 	try {
 		const [tokenPart, timestamp, signature] = token.split(":");
 
@@ -78,14 +99,11 @@ export function verifyCSRFToken(token: string): boolean {
 		const tokenTime = Number.parseInt(timestamp);
 		const now = Date.now();
 		if (now - tokenTime > 3600000) {
-			// 1時間
 			return false;
 		}
 
 		// 署名検証
-		const expectedSignature = createHash("sha256")
-			.update(`${tokenPart}:${timestamp}:${CSRF_SECRET}`)
-			.digest("hex");
+		const expectedSignature = await createSha256Hash(`${tokenPart}:${timestamp}:${CSRF_SECRET}`);
 
 		return signature === expectedSignature;
 	} catch {
@@ -94,14 +112,16 @@ export function verifyCSRFToken(token: string): boolean {
 }
 
 /**
- * リクエストのCSRFトークンをチェック
- * GETリクエストはスキップし、POST等でトークンを検証
+ * リクエストのCSRFトークンをチェックする（Double-submit Cookieパターン）。
+ *
+ * - 安全メソッド（GET / HEAD / OPTIONS）はスキップ
+ * - 開発環境で `CSRF_DEBUG=true` のときはバイパス（本番禁止）
+ * - ヘッダー `x-csrf-token` と Cookie `csrf-token` の両方が必須かつ値一致を要求
  *
  * @param request NextRequest オブジェクト
  * @returns CSRF検証結果
- * @serverOnly このファイルはサーバーサイドでのみ使用可能
  */
-export function checkCSRFToken(request: NextRequest): boolean {
+export async function checkCSRFToken(request: NextRequest): Promise<boolean> {
 	// 安全メソッド（副作用なし）はスキップ。HEAD/OPTIONS（preflight）も含める。
 	if (request.method === "GET" || request.method === "HEAD" || request.method === "OPTIONS") {
 		return true;
@@ -109,12 +129,8 @@ export function checkCSRFToken(request: NextRequest): boolean {
 
 	// 開発環境では緩和（デバッグ用、本番では絶対に有効化しないこと）
 	if (process.env.NODE_ENV === "development" && process.env.CSRF_DEBUG === "true") {
-		logger.error(
-			{
-				nodeEnv: process.env.NODE_ENV,
-				pid: process.pid,
-			},
-			"[CSRF_DEBUG] CSRF protection is DISABLED. NEVER enable this in production.",
+		console.error(
+			`[CSRF_DEBUG] CSRF protection is DISABLED (NODE_ENV=${process.env.NODE_ENV}). NEVER enable this in production.`,
 		);
 		return true;
 	}
@@ -124,37 +140,23 @@ export function checkCSRFToken(request: NextRequest): boolean {
 	const cookieToken = request.cookies.get("csrf-token")?.value;
 
 	if (!headerToken || !cookieToken) {
-		logger.warn(
-			{
-				pathname: request.nextUrl.pathname,
-				method: request.method,
-				hasHeader: Boolean(headerToken),
-				hasCookie: Boolean(cookieToken),
-			},
-			"CSRF token is missing (header or cookie)",
+		console.warn(
+			`CSRF token is missing (header or cookie). pathname=${request.nextUrl.pathname} method=${request.method}`,
 		);
 		return false;
 	}
 
 	if (headerToken !== cookieToken) {
-		logger.warn(
-			{
-				pathname: request.nextUrl.pathname,
-				method: request.method,
-			},
-			"CSRF token mismatch between header and cookie",
+		console.warn(
+			`CSRF token mismatch between header and cookie. pathname=${request.nextUrl.pathname} method=${request.method}`,
 		);
 		return false;
 	}
 
-	const isValid = verifyCSRFToken(headerToken);
+	const isValid = await verifyCSRFToken(headerToken);
 	if (!isValid) {
-		logger.warn(
-			{
-				pathname: request.nextUrl.pathname,
-				method: request.method,
-			},
-			"CSRF token validation failed",
+		console.warn(
+			`CSRF token validation failed. pathname=${request.nextUrl.pathname} method=${request.method}`,
 		);
 	}
 
@@ -162,29 +164,29 @@ export function checkCSRFToken(request: NextRequest): boolean {
 }
 
 /**
- * CSRF保護が必要なパスかどうかを判定
+ * CSRF保護が必要なパスかどうかを判定する。
+ *
+ * - `/api/*` はデフォルトで保護対象（一部例外あり）
+ * - 例外: `/api/stripe/webhook`, `/api/health`, `/api/csrf-token`
+ * - Server Actions は Next.js のビルトインCSRF保護に委ねるため対象外
  *
  * @param pathname リクエストパス
  * @returns CSRF保護が必要かどうか
- * @serverOnly このファイルはサーバーサイドでのみ使用可能
  */
 export function requiresCSRFProtection(pathname: string): boolean {
-	// API routes
 	if (pathname.startsWith("/api/")) {
-		// CSRF保護が不要なAPI（webhook等）
 		const exemptPaths = ["/api/stripe/webhook", "/api/health", "/api/csrf-token"];
 		return !exemptPaths.some((path) => pathname.startsWith(path));
 	}
 
-	// Server Actions（POST/PUT/DELETE等）
-	return true;
+	// Server Actions は Next.js が自動でCSRF保護するため対象外
+	return false;
 }
 
 /**
- * CSRFエラーレスポンス用のヘルパー
+ * CSRF検証失敗時の 403 レスポンス。
  *
  * @returns 403 Forbiddenレスポンス
- * @serverOnly このファイルはサーバーサイドでのみ使用可能
  */
 export function createCSRFErrorResponse(): Response {
 	return new Response("CSRF token validation failed", {
